# Reduce the pool size for asyncFutureCompletionPool 