--- a/.changes/next-release/bugfix-AWSSDKforJavav2-045e79c.json
+++ b/.changes/next-release/bugfix-AWSSDKforJavav2-045e79c.json
@@ -1,0 +1,5 @@
+{
+    "category": "AWS SDK for Java v2", 
+    "type": "bugfix", 
+    "description": "Update the pool size for default async future completion executor service. See [#1251](https://github.com/aws/aws-sdk-java-v2/issues/1251), [#994](https://github.com/aws/aws-sdk-java-v2/issues/994)"
+}