--- a/test/stability-tests/src/it/java/software/amazon/awssdk/stability/tests/transcribestreaming/TranscribeStreamingStabilityTest.java
+++ b/test/stability-tests/src/it/java/software/amazon/awssdk/stability/tests/transcribestreaming/TranscribeStreamingStabilityTest.java
@@ -19,6 +19,7 @@ import java.io.InputStream;
 import java.time.Duration;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.IntFunction;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
@@ -61,6 +62,11 @@ public class TranscribeStreamingStabilityTest extends AwsTestBase {
         if (audioFileInputStream == null) {
             throw new RuntimeException("fail to get the audio input stream");
         }
+    }
+
+    @AfterAll
+    public static void tearDown() {
+        transcribeStreamingClient.close();
     }
 
     @RetryableTest(maxRetries = 3, retryableException = StabilityTestsRetryableException.class)